Refactor to Mixin
Add Mixin to dispatch calls to AnomalibModule based models.
Helps reduce duplicate code
