+++ src/otx/algo/anomaly/padim.py
@@ -7,23 +7,15 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Literal
+from typing import Literal
 
-from anomalib.callbacks.normalization.min_max_normalization import _MinMaxNormalizationCallback
-from anomalib.callbacks.post_processor import _PostProcessorCallback
-from anomalib.models.image import Padim as AnomalibPadim
+from anomalib.models import Padim as AnomalibPadim
 
-from otx.core.model.anomaly import OTXAnomaly
+from otx.core.model.anomaly import AnomalyMixin, OTXAnomaly
 from otx.core.types.task import OTXTaskType
 
-if TYPE_CHECKING:
-    from lightning.pytorch.utilities.types import STEP_OUTPUT
-    from torch.optim.optimizer import Optimizer
 
-    from otx.core.model.anomaly import AnomalyModelInputs, AnomalyModelOutputs
-
-
-class Padim(OTXAnomaly, AnomalibPadim):
+class Padim(AnomalyMixin, AnomalibPadim, OTXAnomaly):
     """OTX Padim model.
 
     Args:
@@ -48,99 +40,10 @@ class Padim(OTXAnomaly, AnomalibPadim):
             OTXTaskType.ANOMALY_SEGMENTATION,
         ] = OTXTaskType.ANOMALY_CLASSIFICATION,
     ) -> None:
-        OTXAnomaly.__init__(self)
-        AnomalibPadim.__init__(
-            self,
+        super().__init__(
             backbone=backbone,
             layers=layers,
             pre_trained=pre_trained,
             n_features=n_features,
         )
-        self.task = task
-
-    def configure_optimizers(self) -> tuple[list[Optimizer], list[Optimizer]] | None:
-        """PADIM doesn't require optimization, therefore returns no optimizers."""
-        return
-
-    def configure_metric(self) -> None:
-        """This does not follow OTX metric configuration."""
-        return
-
-    def on_train_epoch_end(self) -> None:
-        """Callback triggered when the training epoch ends."""
-        return AnomalibPadim.on_train_epoch_end(self)
-
-    def on_validation_start(self) -> None:
-        """Callback triggered when the validation starts."""
-        return AnomalibPadim.on_validation_start(self)
-
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        AnomalibPadim.on_validation_epoch_start(self)
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        AnomalibPadim.on_test_epoch_start(self)
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        AnomalibPadim.on_validation_epoch_end(self)
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        AnomalibPadim.on_test_epoch_end(self)
-
-    def training_step(
-        self,
-        inputs: AnomalyModelInputs,
-        batch_idx: int = 0,
-    ) -> STEP_OUTPUT:
-        """Call training step of the anomalib model."""
-        if not isinstance(inputs, dict):
-            inputs = self._customize_inputs(inputs)
-        return AnomalibPadim.training_step(self, inputs, batch_idx)  # type: ignore[misc]
-
-    def validation_step(
-        self,
-        inputs: AnomalyModelInputs,
-        batch_idx: int = 0,
-    ) -> STEP_OUTPUT:
-        """Call validation step of the anomalib model."""
-        if not isinstance(inputs, dict):
-            inputs = self._customize_inputs(inputs)
-        return AnomalibPadim.validation_step(self, inputs, batch_idx)  # type: ignore[misc]
-
-    def test_step(
-        self,
-        inputs: AnomalyModelInputs,
-        batch_idx: int = 0,
-        **kwargs,
-    ) -> STEP_OUTPUT:
-        """Call test step of the anomalib model."""
-        if not isinstance(inputs, dict):
-            inputs = self._customize_inputs(inputs)
-        return AnomalibPadim.test_step(self, inputs, batch_idx, **kwargs)  # type: ignore[misc]
-
-    def predict_step(
-        self,
-        inputs: AnomalyModelInputs,
-        batch_idx: int = 0,
-        **kwargs,
-    ) -> STEP_OUTPUT:
-        """Call test step of the anomalib model."""
-        if not isinstance(inputs, dict):
-            inputs = self._customize_inputs(inputs)
-        return AnomalibPadim.predict_step(self, inputs, batch_idx, **kwargs)  # type: ignore[misc]
-
-    def forward(
-        self,
-        inputs: AnomalyModelInputs,
-    ) -> AnomalyModelOutputs:
-        """Wrap forward method of the Anomalib model."""
-        outputs = self.validation_step(inputs)
-        # TODO(Ashwin): update forward implementation to comply with other OTX models
-        _PostProcessorCallback._post_process(outputs)  # noqa: SLF001
-        _PostProcessorCallback._compute_scores_and_labels(self, outputs)  # noqa: SLF001
-        _MinMaxNormalizationCallback._normalize_batch(outputs, self)  # noqa: SLF001
-
-        return self._customize_outputs(outputs=outputs, inputs=inputs)
+        self.task = self.get_anomalib_task_type(task)
